framehop::aarch64

Enum UnwindRuleAarch64

pub enum UnwindRuleAarch64 {
    NoOp,
    NoOpIfFirstFrameOtherwiseFp,
    OffsetSp {
        sp_offset_by_16: u16,
    },
    OffsetSpIfFirstFrameOtherwiseStackEndsHere {
        sp_offset_by_16: u16,
    },
    OffsetSpAndRestoreLr {
        sp_offset_by_16: u16,
        lr_storage_offset_from_sp_by_8: i16,
    },
    OffsetSpAndRestoreFpAndLr {
        sp_offset_by_16: u16,
        fp_storage_offset_from_sp_by_8: i16,
        lr_storage_offset_from_sp_by_8: i16,
    },
    UseFramePointer,
    UseFramepointerWithOffsets {
        sp_offset_from_fp_by_8: u16,
        fp_storage_offset_from_fp_by_8: i16,
        lr_storage_offset_from_fp_by_8: i16,
    },
}

Variants

NoOp

(sp, fp, lr) = (sp, fp, lr) Only possible for the first frame. Subsequent frames must get the return address from somewhere other than the lr register to avoid infinite loops.

NoOpIfFirstFrameOtherwiseFp

(sp, fp, lr) = if is_first_frame (sp, fp, lr) else (fp + 16, *fp, *(fp + 8)) Used as a fallback rule.

OffsetSp

(sp, fp, lr) = (sp + 16x, fp, lr) Only possible for the first frame. Subsequent frames must get the return address from somewhere other than the lr register to avoid infinite loops.

Fields

sp_offset_by_16: u16

OffsetSpIfFirstFrameOtherwiseStackEndsHere

(sp, fp, lr) = (sp + 16x, fp, lr) if is_first_frame This rule reflects an ambiguity in DWARF CFI information. When the return address is โundefinedโ because it was omitted, it could mean โsame valueโ, but this is only allowed for the first frame.

Fields

sp_offset_by_16: u16

OffsetSpAndRestoreLr

(sp, fp, lr) = (sp + 16x, fp, *(sp + 8y))

Fields

sp_offset_by_16: u16

lr_storage_offset_from_sp_by_8: i16

OffsetSpAndRestoreFpAndLr

(sp, fp, lr) = (sp + 16x, *(sp + 8y), *(sp + 8z))

Fields

sp_offset_by_16: u16

fp_storage_offset_from_sp_by_8: i16

lr_storage_offset_from_sp_by_8: i16

UseFramePointer

(sp, fp, lr) = (fp + 16, *fp, *(fp + 8))

UseFramepointerWithOffsets

(sp, fp, lr) = (fp + 8x, *(fp + 8y), *(fp + 8z))

Fields

sp_offset_from_fp_by_8: u16

fp_storage_offset_from_fp_by_8: i16

lr_storage_offset_from_fp_by_8: i16