Enum framehop::x86_64::UnwindRuleX86_64

pub enum UnwindRuleX86_64 {
    EndOfStack,
    JustReturn,
    JustReturnIfFirstFrameOtherwiseFp,
    OffsetSp {
        sp_offset_by_8: u16,
    },
    OffsetSpAndRestoreBp {
        sp_offset_by_8: u16,
        bp_storage_offset_from_sp_by_8: i16,
    },
    UseFramePointer,
    OffsetSpAndPopRegisters {
        sp_offset_by_8: u16,
        register_count: u8,
        encoded_registers_to_pop: u16,
    },
}

For all of these: return address is *(new_sp - 8)

Variants

EndOfStack

JustReturn

(sp, bp) = (sp + 8, bp)

JustReturnIfFirstFrameOtherwiseFp

(sp, bp) = if is_first_frame (sp + 8, bp) else (bp + 16, *bp)

OffsetSp

Fields

sp_offset_by_8: u16

(sp, bp) = (sp + 8x, bp)

OffsetSpAndRestoreBp

Fields

sp_offset_by_8: u16

bp_storage_offset_from_sp_by_8: i16

(sp, bp) = (sp + 8x, *(sp + 8y))

UseFramePointer

(sp, bp) = (bp + 16, *bp)

OffsetSpAndPopRegisters

Fields

sp_offset_by_8: u16

The additional stack pointer offset to undo before popping the registers, divided by 8 bytes.

register_count: u8

The number of registers to pop from the stack.

encoded_registers_to_pop: u16

An encoded ordering of the callee-save registers to pop from the stack, see register_ordering.

(sp, …) = (sp + 8 * (offset + register count), … popped according to encoded ordering) This supports the common case of pushed callee-saved registers followed by a stack allocation. Up to 8 registers can be stored, which covers all callee-saved registers (aside from RSP which is implicit).

The registers are stored in a separate compressed ordering to facilitate restoring register values if desired. If not for this we could simply store the total offset.

Implementations

impl UnwindRuleX86_64

pub fn for_sequence_of_offset_or_pop<I, T>(iter: I) -> Option<Self>
where I: Iterator<Item = T>, T: Into<OffsetOrPop>,

Get the rule which represents the given operations, if possible.